add Bootstrap styling, reserve button, and initial delete button implementation

+++ cinema-reservation-app/src/main/java/com/cinemareservation/repository/FilmRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface FilmRepository  extends JpaRepository<Film, Integer> {
+public interface FilmRepository  extends JpaRepository<Film, Long> {
 
 }
